Fix remaining dialog import paths after reorganization
